+++ package.json
@@ -46,7 +46,7 @@
     "autoprefixer": "^10.4.7",
     "clsx": "^1.2.1",
     "daisyui": "^2.19.0",
-    "eslint": "^7.32.0 || ^8.2.0",
+    "eslint": "^8.2.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.25.3",
