+++ src/component/atom/ModalActions/ModalActions.tsx
@@ -1,12 +1,8 @@
 import clsx from 'clsx'
 import React from 'react'
 
-import { Button, IButtonProps } from '../button/Button'
-
-interface IModalActionsProps {
-  actions: IButtonProps[]
-  className?: string
-}
+import { IButtonProps, IModalActionsProps } from '../../../types'
+import { Button } from '../button/Button'
 
 export const ModalActions = ({ className = '', actions }: IModalActionsProps) => {
   const classes = clsx(['modal-action border-t p-0 m-0 px-3 py-2', className])
