+++ src/component/atom/icon/Icon.tsx
@@ -1,21 +1,7 @@
 import React from 'react'
 
+import { IconProps } from '../../../types'
 import { isString } from '../../../utils/common/common'
-
-type OUTLINE_IDENTIFIER = 'circular' | 'rectangle'
-
-export interface IconProps {
-  source: string | (() => any)
-  onClick?: (_?: any) => void | unknown
-  style?: object
-  outlined?: OUTLINE_IDENTIFIER
-  size?: number
-  height?: number
-  iconClass?: string
-  className?: string
-  iconColor?: string
-  isSvg?: boolean
-}
 
 export const Icon = ({
   source,
