+++ src/component/atom/index.ts
@@ -1,4 +1,3 @@
-export type { IButtonProps } from './button/Button'
 export { Button } from './button/Button'
 export { Icon } from './icon/Icon'
 export * from './ModalActions/ModalActions'
