+++ src/component/organism/GlobalModalComponent/index.tsx
@@ -1,31 +1,7 @@
 import React, { forwardRef } from 'react'
 
-import { IButtonProps } from '../../atom'
+import { GlobalModalComponentProps } from '../../../types'
 import { Modal } from '../../molecules'
-
-export type GlobalModalComponentProps = {
-  component?: React.FC<any>
-  props?: { [key: string]: unknown }
-  onClose?: () => void
-  closeModal?: () => void
-  isOpen?: boolean
-  responsive?: boolean
-  children?: any
-  className?: string
-  modalSize?: 'xs' | 'sm' | 'md' | 'lg'
-  isCloseable?: boolean
-  closeButtonClassName?: string
-  title?: string
-  hideHeader?: boolean
-  headerComponent?: React.FC<any>
-  headerClassName?: string
-  actions?: IButtonProps[]
-  actionClassName?: string
-  contentClassName?: string
-  hideCloseIcon?: boolean
-  height?: number
-  width?: number
-}
 
 export const GlobalModalComponent = forwardRef(
   (
