+++ src/types/index.ts
@@ -1,1 +1,4 @@
+export * from './Button'
+export * from './Icon'
+export * from './Modal'
 export * from './Routes'
